requirements.txt
openai
speechrecognition
pyaudio
python-dotenv
